separated into files & simplified movement handle
